Split tests in packages
