+++ packages/strapi-plugin-content-manager/jest.config.js
@@ -1,7 +1,9 @@
 module.exports = {
   name: 'content-manager',
   displayName: 'Content Manager',
+  testMatch: ['test/?(*.)+(test).js'],
   coveragePathIgnorePatterns: [
+    '<rootDir>/admin/',
     '<rootDir>/dist/',
     '<rootDir>/node_modules/',
     '<rootDir>/out-tsc/',
